requirements.txt
pyTelegramBotAPI
google-genai
SpeechRecognition
pydub
pillow
pytesseract
python-dotenv